# refactor!: replace `PartSet` parts slice with buffer