--- a/consensus/replay_test.go
+++ b/consensus/replay_test.go
@@ -1071,7 +1071,7 @@ func makeBlockchainFromWAL(wal WAL) ([]*types.Block, []*types.ExtendedCommit, er
 				height++
 			}
 		case *types.PartSetHeader:
-			thisBlockParts = types.NewPartSetFromHeader(*p)
+			thisBlockParts = types.NewPartSetFromHeader(*p, types.BlockPartSizeBytes)
 		case *types.Part:
 			_, err := thisBlockParts.AddPart(p)
 			if err != nil {