--- a/consensus/replay_test.go
+++ b/consensus/replay_test.go
@@ -1046,10 +1046,7 @@ func makeBlockchainFromWAL(wal WAL) ([]*types.Block, []*types.ExtendedCommit, er
 			// if its not the first one, we have a full block
 			if thisBlockParts != nil {
 				pbb := new(cmtproto.Block)
-				bz, err := io.ReadAll(thisBlockParts.GetReader())
-				if err != nil {
-					panic(err)
-				}
+				bz := thisBlockParts.GetBytes()
 				err = proto.Unmarshal(bz, pbb)
 				if err != nil {
 					panic(err)
@@ -1089,10 +1086,7 @@ func makeBlockchainFromWAL(wal WAL) ([]*types.Block, []*types.ExtendedCommit, er
 		}
 	}
 	// grab the last block too
-	bz, err := io.ReadAll(thisBlockParts.GetReader())
-	if err != nil {
-		panic(err)
-	}
+	bz := thisBlockParts.GetBytes()
 	pbb := new(cmtproto.Block)
 	err = proto.Unmarshal(bz, pbb)
 	if err != nil {