## Summary
Enhanced `PartSet` by utilizing a single preallocated buffer for part storage, reducing memory overhead. Updated related methods, removing part array access in favor of buffer management.

The biggest difference is in reading - because we can return entire buffer, it's 0 overhead., while previous implementation using io.ReadAll and custom Reader was way slower.
In artificial benchmarks, inserting parts into PartSet takes similar time, but some allocations are saved - hopefully this could be more significant in real world.

## Results from my machine
### Before changes
```
$ go test -bench BenchmarkPartSetRoundTrip -benchtime 1s
goos: linux
goarch: amd64
pkg: github.com/cometbft/cometbft/types
cpu: Intel(R) Core(TM) i7-10875H CPU @ 2.30GHz
BenchmarkPartSetRoundTrip/dataSize=1000,partSize=256/write-16   	  133821	      8993 ns/op	    2238 B/op	      21 allocs/op
BenchmarkPartSetRoundTrip/dataSize=1000,partSize=256/read-16    	 1319215	       919.1 ns/op	    3056 B/op	       8 allocs/op
BenchmarkPartSetRoundTrip/dataSize=130857600,partSize=65536/write-16         	       3	 354859913 ns/op	148265408 B/op	   27931 allocs/op
BenchmarkPartSetRoundTrip/dataSize=130857600,partSize=65536/read-16          	      12	  95687679 ns/op	769322953 B/op	    2048 allocs/op
BenchmarkPartSetRoundTrip/dataSize=523430400,partSize=65536/write-16         	       1	1460525033 ns/op	593392160 B/op	  127686 allocs/op
BenchmarkPartSetRoundTrip/dataSize=523430400,partSize=65536/read-16          	       3	 386639023 ns/op	2937041840 B/op	    8050 allocs/op
PASS
ok  	github.com/cometbft/cometbft/types	49.068s
```
### After changes
```
$ go test -bench BenchmarkPartSetRoundTrip -benchtime 1s
goos: linux
goarch: amd64
pkg: github.com/cometbft/cometbft/types
cpu: Intel(R) Core(TM) i7-10875H CPU @ 2.30GHz
BenchmarkPartSetRoundTrip/dataSize=1000,partSize=256/write-16   	  133736	      8643 ns/op	    1790 B/op	      17 allocs/op
BenchmarkPartSetRoundTrip/dataSize=1000,partSize=256/read-16    	38419419	        30.62 ns/op	       0 B/op	       0 allocs/op
BenchmarkPartSetRoundTrip/dataSize=130857600,partSize=65536/write-16         	       3	 369702647 ns/op	148041744 B/op	   25934 allocs/op
BenchmarkPartSetRoundTrip/dataSize=130857600,partSize=65536/read-16          	48963622	        24.23 ns/op	       0 B/op	       0 allocs/op
BenchmarkPartSetRoundTrip/dataSize=523430400,partSize=65536/write-16         	       1	1463117711 ns/op	592497568 B/op	  119697 allocs/op
BenchmarkPartSetRoundTrip/dataSize=523430400,partSize=65536/read-16          	44423292	        25.45 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/cometbft/cometbft/types	45.972s
```

Resolves #2321