@rach-id 2025-08-21
Since this change is critical, I asked review from copilot to make sure all byte operations are in order.